# shared/api-spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Review Anything API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/notes/section/{section_name}": {
      "get": {
        "tags": [
          "notes"
        ],
        "summary": "Get Notes By Section",
        "description": "获取某板块下的所有笔记",
        "operationId": "get_notes_by_section_api_notes_section__section_name__get",
        "parameters": [
          {
            "name": "section_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Section Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/review/next": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Get Next Review",
        "description": "获取下一个要复习的条目",
        "operationId": "get_next_review_api_review_next_get",
        "parameters": [
          {
            "name": "section",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Section"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/review/batch": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Get Batch Review",
        "description": "批量获取复习条目",
        "operationId": "get_batch_review_api_review_batch_get",
        "parameters": [
          {
            "name": "section",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Section"
            }
          },
          {
            "name": "count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Count"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/review/{item_id}/action": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Review Action",
        "description": "处理复习动作",
        "operationId": "review_action_api_review__item_id__action_post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewAction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sections/": {
      "get": {
        "tags": [
          "sections"
        ],
        "summary": "List Sections",
        "description": "获取所有板块（包括手动创建的）",
        "operationId": "list_sections_api_sections__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "sections"
        ],
        "summary": "Create Section",
        "description": "手动创建板块",
        "operationId": "create_section_api_sections__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SectionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sections/rename": {
      "post": {
        "tags": [
          "sections"
        ],
        "summary": "Rename Section",
        "description": "重命名板块（会同时更新 Section 表和 Note 表中的板块名）",
        "operationId": "rename_section_api_sections_rename_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SectionRename"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sections/{section_name}": {
      "delete": {
        "tags": [
          "sections"
        ],
        "summary": "Delete Section",
        "description": "删除板块及其下的所有笔记、分块和复习项",
        "operationId": "delete_section_api_sections__section_name__delete",
        "parameters": [
          {
            "name": "section_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Section Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ReviewAction": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "ReviewAction"
      },
      "SectionCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "SectionCreate"
      },
      "SectionRename": {
        "properties": {
          "old_name": {
            "type": "string",
            "title": "Old Name"
          },
          "new_name": {
            "type": "string",
            "title": "New Name"
          }
        },
        "type": "object",
        "required": [
          "old_name",
          "new_name"
        ],
        "title": "SectionRename"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}